AutoComplete Location feature added to Form and SearchMemories
