+++ memories-app/package.json
@@ -17,6 +17,7 @@
     "react-bootstrap": "^2.6.0",
     "react-bootstrap-validation": "^0.1.11",
     "react-dom": "^18.2.0",
+    "react-google-autocomplete": "^2.7.0",
     "react-icons": "^4.6.0",
     "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
