+++ memories-app/src/App.js
@@ -1,15 +1,13 @@
 import React from "react";
 import "./App.css";
 import Header from "./Components/Header";
-import SearchBar from "./Components/SearchBar/SearchBar";
-import FloatingButton from "./Components/FloatingButton/FloatingButton";
 import Footer from "./Components/Footer/Footer";
 import { BrowserRouter } from "react-router-dom";
 import { Route, Routes, Navigate } from "react-router-dom";
 import HomePage from "./Pages/Home/HomePage";
 import InputMemory from "./Pages/InputMemory/InputMemory";
 import MemoryCard from "./Components/MemoryCards";
-import SearchCard from "./Components/SearchCard";
+import SearchCard from "./Components/SearchBar/SearchCard";
 
 function App() {
   return (
