+++ memories-app/src/Components/SearchBar/SearchCard.css
@@ -1,0 +1,48 @@
+/* Searchbar Styling */
+form {
+  text-align: center;
+  padding: auto;
+  width: 50%;
+  border: 3px solid #0A5C8F;
+  border-radius: 20px;
+  padding: 9px 4px 9px 40px;
+  background: transparent url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='16' height='16' class='bi bi-search' viewBox='0 0 16 16'%3E%3Cpath d='M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z'%3E%3C/path%3E%3C/svg%3E") no-repeat 13px center;
+}
+input {
+  width: 50%;
+  border: 3px solid #0A5C8F;
+  border-radius: 20px;
+  padding: 9px 4px 9px 40px;
+  background: transparent url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='16' height='16' class='bi bi-search' viewBox='0 0 16 16'%3E%3Cpath d='M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z'%3E%3C/path%3E%3C/svg%3E") no-repeat 13px center;
+}
+.icon {
+  text-align: center;
+
+  height: 40px;
+  width: 40px; 
+}
+.search-card{
+  justify-content: center;
+  text-align: center;
+}
+.form{
+  text-align: center;
+}
+
+.ui.container {
+  margin-left: 0 !important;
+  margin-right: 0 !important;
+
+}
+
+@media (max-width: 800px) {
+  .ui.container {
+    flex-direction: column;
+    align-items: center;
+  }
+}
+
+.cardDisplay{
+  justify-content: center;
+  text-align: center;
+}
