+++ memories-app/src/Components/Viewcard/ViewCard.css
