+++ memories-app/src/Pages/Home/HomePage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import FloatingButton from '../../Components/FloatingButton/FloatingButton'
-import SearchBar from '../../Components/SearchBar/SearchBar';
+//import SearchBar from '../../Components/SearchBar/SearchBar';
 import About from '../../Components/About';
 import Welcome from '../../Components/Welcome';
